requirements.txt
streamlit
pandas
numpy
torch
transformers
pillow